Refactor the signature of the RemoteUpdater class
